+++ modules/generator/processor/spanmetrics/spanmetrics.go
@@ -58,14 +58,18 @@ func New(cfg Config, registry registry.Registry) gen.Processor {
 		labels = append(labels, sanitizeLabelNameWithCollisions(d))
 	}
 
-	return &Processor{
-		Cfg:                        cfg,
-		registry:                   registry,
-		spanMetricsCallsTotal:      registry.NewCounter(metricCallsTotal, labels),
-		spanMetricsDurationSeconds: registry.NewHistogram(metricDurationSeconds, labels, cfg.HistogramBuckets),
-		spanMetricsSizeTotal:       registry.NewCounter(metricSizeTotal, labels),
-		now:                        time.Now,
+	p := &Processor{}
+	if cfg.Subprocessors["Latency"] {
+		p.spanMetricsDurationSeconds = registry.NewHistogram(metricDurationSeconds, labels, cfg.HistogramBuckets)
 	}
+	if cfg.Subprocessors["Count"] {
+		p.spanMetricsCallsTotal = registry.NewCounter(metricCallsTotal, labels)
+	}
+	p.Cfg = cfg
+	p.registry = registry
+	p.spanMetricsSizeTotal = registry.NewCounter(metricSizeTotal, labels)
+	p.now = time.Now
+	return p
 }
 
 func (p *Processor) Name() string {
@@ -122,9 +126,15 @@ func (p *Processor) aggregateMetricsForSpan(svcName string, rs *v1.Resource, spa
 
 	registryLabelValues := p.registry.NewLabelValues(labelValues)
 
-	p.spanMetricsCallsTotal.Inc(registryLabelValues, 1)
+	if p.Cfg.Subprocessors["Count"] {
+		p.spanMetricsCallsTotal.Inc(registryLabelValues, 1)
+	}
+
 	p.spanMetricsSizeTotal.Inc(registryLabelValues, float64(span.Size()))
-	p.spanMetricsDurationSeconds.ObserveWithExemplar(registryLabelValues, latencySeconds, tempo_util.TraceIDToHexString(span.TraceId))
+
+	if p.Cfg.Subprocessors["Latency"] {
+		p.spanMetricsDurationSeconds.ObserveWithExemplar(registryLabelValues, latencySeconds, tempo_util.TraceIDToHexString(span.TraceId))
+	}
 }
 
 func sanitizeLabelNameWithCollisions(name string) string {